+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - 2019-11-06
+### Added
+- Logs for write methods: 
+	- insert
+	- multiInsert
+	- update
+	- save
+	- multiSave
+	- remove
+	- multiRemove
+- `@janiscommerce/log` package
+- `lib/utils.js` module
+
 ## [3.2.0] - 2019-11-06
 ### Added
 - Distinct method implemented with driver support check